日期类型统一修改为 long

+++ jobs-core/src/main/java/com/baomidou/jobs/core/executor/IJobsExecutor.java
@@ -24,7 +24,7 @@ public interface IJobsExecutor {
      * @param jobId
      * @return
      */
-    JobsResponse<String> idleBeat(int jobId);
+    JobsResponse<String> idleBeat(Long jobId);
 
     /**
      * kill
@@ -32,7 +32,7 @@ public interface IJobsExecutor {
      * @param jobId
      * @return
      */
-    JobsResponse<String> kill(int jobId);
+    JobsResponse<String> kill(Long jobId);
 
     /**
      * run
