+++ jobs-spring-boot-starter/src/main/java/com/baomidou/jobs/starter/router/strategy/ExecutorRouteLFU.java
@@ -16,10 +16,10 @@ import java.util.concurrent.ConcurrentHashMap;
  */
 public class ExecutorRouteLFU implements IExecutorRouter {
 
-    private static ConcurrentHashMap<Integer, HashMap<String, Integer>> jobLfuMap = new ConcurrentHashMap<Integer, HashMap<String, Integer>>();
+    private static ConcurrentHashMap<Long, HashMap<String, Integer>> jobLfuMap = new ConcurrentHashMap<Long, HashMap<String, Integer>>();
     private static long CACHE_VALID_TIME = 0;
 
-    public String route(int jobId, List<String> addressList) {
+    public String route(Long jobId, List<String> addressList) {
 
         // cache clear
         if (System.currentTimeMillis() > CACHE_VALID_TIME) {
